Add/Remove automatic help creation
Created an rudimentary automatic about_ help creator, but it looked bad.
Code is still in there but all Remarked out.  Added some brief pauses in
between Add-Content because on my laptop it was moving on to the next
Add-Content before the previous one was finished causing the second one
to fail.